Use Red Hat IT's custom sonar scanner container image
The IT team has built a custom image which already has the keystore
generated with Red Hat's IT custom certificate in it. This simplifies
the scan process since we only have to mount the code in the container
and run the scanner, instead of having to create the keystore ourselves
and having to download the scanner every time the pipeline ran.